refactoring code using context
refactoring code using context
